+++ env.example
@@ -1,6 +1,1 @@
-# URL for your Directus instance
-# Whether that's your local (http://localhost:8055) or Directus Cloud (https://your-instance.directus.app)
 DIRECTUS_URL="http://localhost:8055"
-
-# Used for seeding database with /setup/seed script
-DIRECTUS_ADMIN_TOKEN="xxxxxxxxxx"
