+++ nuxt.config.ts
@@ -1,3 +1,5 @@
+import { defineNuxtConfig } from 'nuxt'
+
 // https://v3.nuxtjs.org/docs/directory-structure/nuxt.config
 export default defineNuxtConfig({
   css: ['~~/assets/css/tailwind.css'],
