+++ package.json
@@ -7,9 +7,6 @@
         "allupdate": "yarn upgrade-interactive --latest"
     },
     "devDependencies": {
-        "@directus/sdk": "^10.1.4",
-        "@headlessui/vue": "^1.7.1",
-        "@heroicons/vue": "^2.0.11",
         "@pinia/nuxt": "^0.4.2",
         "@tailwindcss/aspect-ratio": "0.4.2",
         "@tailwindcss/forms": "0.5.3",
@@ -18,14 +15,18 @@
         "@vueuse/core": "^9.2.0",
         "@vueuse/nuxt": "^9.2.0",
         "autoprefixer": "10.4.9",
+        "eslint": "^8.23.1",
         "eslint-config-prettier": "^8.5.0",
         "eslint-plugin-prettier": "^4.2.1",
-        "eslint": "^8.23.1",
         "nuxt": "^3.0.0-rc.11",
-        "pinia": "^2.0.22",
         "postcss": "^8.4.16",
         "prettier": "^2.7.1",
         "tailwindcss": "^3.1.8"
     },
-    "dependencies": {}
+    "dependencies": {
+        "@directus/sdk": "^10.1.4",
+        "@headlessui/vue": "^1.7.1",
+        "@heroicons/vue": "^2.0.11",
+        "pinia": "^2.0.22"
+    }
 }
