+++ plugins/directus.ts
@@ -1,11 +1,12 @@
 import { BaseStorage, Directus } from '@directus/sdk'
 import { useAuth } from '~~/store/auth'
-import { readRawCookies, writeClientCookie, writeServerCookie } from './cookies'
 
 // Make sure you review the Directus SDK documentation for more information
 // https://docs.directus.io/reference/sdk.html
+
 export default defineNuxtPlugin(async (nuxtApp) => {
   const { directusUrl } = useRuntimeConfig()
+
   const auth = useAuth()
 
   // Create a new storage class to use with the SDK
@@ -14,26 +15,18 @@ export default defineNuxtPlugin(async (nuxtApp) => {
     deletedKeys = new Set<string>()
     get(key) {
       if (this.deletedKeys.has(key)) return null
-      const cookies = readRawCookies(
-        process.server ? nuxtApp.ssrContext.req : null
-      )
-      return cookies[key]
+      const cookie = useCookie(key)
+      return cookie.value
     }
     set(key, value) {
       this.deletedKeys.delete(key)
-      if (process.client) {
-        writeClientCookie(key, value)
-      } else if (process.server) {
-        writeServerCookie(nuxtApp.ssrContext.res, key, value)
-      }
+      const cookie = useCookie(key)
+      return (cookie.value = value)
     }
     delete(key) {
       this.deletedKeys.add(key)
-      if (process.client) {
-        writeClientCookie(key, null)
-      } else if (process.server) {
-        writeServerCookie(nuxtApp.ssrContext.res, key, null)
-      }
+      const cookie = useCookie(key)
+      return (cookie.value = null)
     }
   }
 
@@ -45,13 +38,13 @@ export default defineNuxtPlugin(async (nuxtApp) => {
     },
   })
 
-  //   // Inject the SDK into the Nuxt app
+  // Inject the SDK into the Nuxt app
   nuxtApp.provide('directus', directus)
 
-  const side = process.server ? 'server' : 'client'
   const token = await directus.auth.token
+  const side = process.server ? 'server' : 'client'
 
-  //   If there's a token but we don't have a user, fetch the user
+  // If there's a token but we don't have a user, fetch the user
   if (!auth.isLoggedIn && token) {
     console.log('Token found, fetching user from ' + side)
     console.log('Token is', token)
