--- setup/collections.json
@@ -1,185 +1,0 @@
-[
-    {
-        "collection": "page_builder",
-        "meta": {
-            "collection": "page_builder",
-            "icon": "import_export",
-            "note": null,
-            "display_template": null,
-            "hidden": true,
-            "singleton": false,
-            "translations": null,
-            "archive_field": null,
-            "archive_app_filter": true,
-            "archive_value": null,
-            "unarchive_value": null,
-            "sort_field": null,
-            "accountability": "all",
-            "color": null,
-            "item_duplication_fields": null,
-            "sort": 3,
-            "group": null,
-            "collapse": "open"
-        },
-        "schema": {
-            "name": "page_builder",
-            "sql": "CREATE TABLE \"page_builder\" (`id` integer PRIMARY KEY AUTOINCREMENT NOT NULL, `pages_id` char(36) NULL, `item` varchar(255) NULL, `sort` integer NULL, `collection` varchar(255) NULL, CONSTRAINT `page_builder_pages_id_foreign` FOREIGN KEY (`pages_id`) REFERENCES `pages` (`id`) ON DELETE SET NULL)"
-        }
-    },
-    {
-        "collection": "page_builder_content",
-        "meta": {
-            "collection": "page_builder_content",
-            "icon": null,
-            "note": null,
-            "display_template": null,
-            "hidden": true,
-            "singleton": false,
-            "translations": [
-                {
-                    "language": "en-US",
-                    "translation": "Content",
-                    "singular": "Content",
-                    "plural": "Content"
-                }
-            ],
-            "archive_field": null,
-            "archive_app_filter": true,
-            "archive_value": null,
-            "unarchive_value": null,
-            "sort_field": null,
-            "accountability": "all",
-            "color": null,
-            "item_duplication_fields": null,
-            "sort": 3,
-            "group": "page_builder",
-            "collapse": "open"
-        },
-        "schema": {
-            "name": "page_builder_content",
-            "sql": "CREATE TABLE \"page_builder_content\" (`id` char(36) NOT NULL, `headline` varchar(255) null, `content` text null, PRIMARY KEY (`id`))"
-        }
-    },
-    {
-        "collection": "page_builder_features",
-        "meta": {
-            "collection": "page_builder_features",
-            "icon": null,
-            "note": null,
-            "display_template": null,
-            "hidden": true,
-            "singleton": false,
-            "translations": [
-                {
-                    "language": "en-US",
-                    "translation": "Features",
-                    "singular": "Features",
-                    "plural": "Features"
-                }
-            ],
-            "archive_field": null,
-            "archive_app_filter": true,
-            "archive_value": null,
-            "unarchive_value": null,
-            "sort_field": null,
-            "accountability": "all",
-            "color": null,
-            "item_duplication_fields": null,
-            "sort": 2,
-            "group": "page_builder",
-            "collapse": "open"
-        },
-        "schema": {
-            "name": "page_builder_features",
-            "sql": "CREATE TABLE \"page_builder_features\" (`id` char(36) NOT NULL, `headline` varchar(255) null, `features` json null, PRIMARY KEY (`id`))"
-        }
-    },
-    {
-        "collection": "page_builder_hero",
-        "meta": {
-            "collection": "page_builder_hero",
-            "icon": null,
-            "note": null,
-            "display_template": "{{headline}}",
-            "hidden": true,
-            "singleton": false,
-            "translations": [
-                {
-                    "language": "en-US",
-                    "translation": "Hero",
-                    "singular": "hero",
-                    "plural": "hero"
-                }
-            ],
-            "archive_field": null,
-            "archive_app_filter": true,
-            "archive_value": null,
-            "unarchive_value": null,
-            "sort_field": null,
-            "accountability": "all",
-            "color": null,
-            "item_duplication_fields": null,
-            "sort": 1,
-            "group": "page_builder",
-            "collapse": "open"
-        },
-        "schema": {
-            "name": "page_builder_hero",
-            "sql": "CREATE TABLE \"page_builder_hero\" (`id` char(36) NOT NULL, `headline` varchar(255) NULL, `content` text NULL, `image` char(36) NULL, `buttons` json NULL, PRIMARY KEY (`id`), CONSTRAINT `page_builder_hero_image_foreign` FOREIGN KEY (`image`) REFERENCES `directus_files` (`id`) ON DELETE SET NULL)"
-        }
-    },
-    {
-        "collection": "pages",
-        "meta": {
-            "collection": "pages",
-            "icon": null,
-            "note": null,
-            "display_template": null,
-            "hidden": false,
-            "singleton": false,
-            "translations": null,
-            "archive_field": "status",
-            "archive_app_filter": true,
-            "archive_value": "archived",
-            "unarchive_value": "draft",
-            "sort_field": null,
-            "accountability": "all",
-            "color": null,
-            "item_duplication_fields": null,
-            "sort": 1,
-            "group": null,
-            "collapse": "open"
-        },
-        "schema": {
-            "name": "pages",
-            "sql": "CREATE TABLE \"pages\" (`id` char(36) NOT NULL, `status` varchar(255) NOT NULL DEFAULT 'draft', `date_created` datetime NULL, `user_updated` char(36) NULL, `date_updated` datetime NULL, `user_created` char(36) NULL, `title` varchar(255) NULL, `slug` varchar(255) NOT NULL DEFAULT null, `image` char(36) NULL, PRIMARY KEY (`id`), CONSTRAINT `pages_user_updated_foreign` FOREIGN KEY (`user_updated`) REFERENCES `directus_users` (`id`), CONSTRAINT `pages_user_created_foreign` FOREIGN KEY (`user_created`) REFERENCES `directus_users` (`id`) ON DELETE SET NULL, CONSTRAINT `pages_image_foreign` FOREIGN KEY (`image`) REFERENCES `directus_files` (`id`) ON DELETE SET NULL)"
-        }
-    },
-    {
-        "collection": "posts",
-        "meta": {
-            "collection": "posts",
-            "icon": null,
-            "note": null,
-            "display_template": null,
-            "hidden": false,
-            "singleton": false,
-            "translations": null,
-            "archive_field": "status",
-            "archive_app_filter": true,
-            "archive_value": "archived",
-            "unarchive_value": "draft",
-            "sort_field": "sort",
-            "accountability": "all",
-            "color": null,
-            "item_duplication_fields": null,
-            "sort": 2,
-            "group": null,
-            "collapse": "open"
-        },
-        "schema": {
-            "name": "posts",
-            "sql": "CREATE TABLE \"posts\" (`id` char(36) NOT NULL, `status` varchar(255) NOT NULL DEFAULT 'draft', `sort` integer NULL, `user_created` char(36) NULL, `date_created` datetime NULL, `user_updated` char(36) NULL, `date_updated` datetime NULL, `slug` varchar(255) NOT NULL DEFAULT null, `title` varchar(255) NULL DEFAULT null, `content` text NULL DEFAULT null, `image` char(36) NULL, PRIMARY KEY (`id`), CONSTRAINT `posts_user_created_foreign` FOREIGN KEY (`user_created`) REFERENCES `directus_users` (`id`), CONSTRAINT `posts_user_updated_foreign` FOREIGN KEY (`user_updated`) REFERENCES `directus_users` (`id`), CONSTRAINT `posts_image_foreign` FOREIGN KEY (`image`) REFERENCES `directus_files` (`id`) ON DELETE SET NULL)"
-        }
-    }
-]
