--- setup/items.json
@@ -1,8 +1,0 @@
-{
-    "pages": [],
-    "posts": [],
-    "page_builder": [],
-    "page_builder_hero": [],
-    "page_builder_content": [],
-    "page_builder_features": []
-}
